refactor(group): remove GroupService interface
